New FPL Take 3
Clean up the FHIRDefinitions objects to only contain what's actually needed.
