+++ src/fhirdefs/index.ts
@@ -1,3 +1,2 @@
-export * from './BaseFHIRDefinitions';
 export * from './FHIRDefinitions';
 export * from './R5DefsForR4';
